challenges/Day 6 Task/fraud-agent/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-silero
livekit-murf
python-dotenv
pymongo
